+++ src/main/java/org/uiop/easyplacefix/IisSimpleHitPos.java
@@ -1,6 +1,0 @@
-package org.uiop.easyplacefix;
-
-public interface IisSimpleHitPos {
-    default void setSimpleHitPos() {
-    }
-}