+++ src/main/java/org/uiop/easyplacefix/Mixin/packet/MixinPlayerMoveC2SPacket.java
@@ -7,7 +7,8 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import static com.tick_ins.tick.TickThread.notChangPlayerLook;
+import com.tick_ins.tick.TickThread;
+
 import static com.tick_ins.tick.TickThread.pitchLock;
 import static com.tick_ins.tick.TickThread.yawLock;
 
@@ -28,7 +29,7 @@ public interface MixinPlayerMoveC2SPacket {
     class Full {
         @Inject(method = "write", at = @At("HEAD"))
         private void lockLook(FriendlyByteBuf buf, CallbackInfo ci) {
-            if (notChangPlayerLook) {
+            if (TickThread.isLookLocked()) {
                 ((MixinPlayerMoveC2SPacket) this).setYRot(yawLock);
                 ((MixinPlayerMoveC2SPacket) this).setXRot(pitchLock);
             }
@@ -39,7 +40,7 @@ public interface MixinPlayerMoveC2SPacket {
     class LookAndOnGround {
         @Inject(method = "write", at = @At("HEAD"))
         private void lockLook(FriendlyByteBuf buf, CallbackInfo ci) {
-            if (notChangPlayerLook) {
+            if (TickThread.isLookLocked()) {
                 ((MixinPlayerMoveC2SPacket) this).setYRot(yawLock);
                 ((MixinPlayerMoveC2SPacket) this).setXRot(pitchLock);
             }